renaming proxy module to bungeecord
