--- api/api-proxy/build.gradle
@@ -1,12 +1,0 @@
-plugins {
-    id 'java'
-}
-
-group 'net.suqatri.redicloud.api.proxy'
-
-dependencies {
-    compileOnly('net.md-5:bungeecord-api:1.19-R0.1-SNAPSHOT')
-    implementation project(':api:api-default-impl')
-    implementation project(':api:api')
-    implementation project(':commons')
-}
