--- api/api-proxy/src/main/java/net/suqatri/redicloud/api/proxy/ProxyDefaultCloudAPI.java
@@ -1,22 +1,0 @@
-package net.suqatri.redicloud.api.proxy;
-
-import net.suqatri.redicloud.api.impl.CloudDefaultAPIImpl;
-import net.suqatri.redicloud.api.impl.redis.bucket.RBucketObject;
-import net.suqatri.redicloud.api.impl.service.CloudService;
-import net.suqatri.redicloud.api.player.ICloudPlayer;
-import net.suqatri.redicloud.api.player.IPlayerBridge;
-import net.suqatri.redicloud.api.proxy.player.ProxyPlayerBridge;
-import net.suqatri.redicloud.api.redis.bucket.IRBucketHolder;
-import net.suqatri.redicloud.api.utils.ApplicationType;
-
-public abstract class ProxyDefaultCloudAPI extends CloudDefaultAPIImpl<CloudService> {
-
-    public ProxyDefaultCloudAPI() {
-        super(ApplicationType.SERVICE_PROXY);
-    }
-
-    @Override
-    public IPlayerBridge createBridge(IRBucketHolder<ICloudPlayer> playerHolder) {
-        return new ProxyPlayerBridge(playerHolder);
-    }
-}
