+++ plugins/plugin-bungeecord/src/main/java/net/suqatri/redicloud/plugin/bungeecord/BungeeCordCloudPlugin.java
@@ -1,0 +1,29 @@
+package net.suqatri.redicloud.plugin.bungeecord;
+
+import net.md_5.bungee.api.ProxyServer;
+import net.md_5.bungee.api.plugin.Plugin;
+import net.suqatri.redicloud.api.service.ServiceState;
+
+public class BungeeCordCloudPlugin extends Plugin {
+
+    private BungeeCordCloudAPI cloudAPI;
+
+    @Override
+    public void onLoad() {
+        cloudAPI = new BungeeCordCloudAPI(this);
+    }
+
+    @Override
+    public void onEnable() {
+        if (cloudAPI == null) throw new IllegalStateException("CloudAPI is not initialized yet!");
+        cloudAPI.registerStartedService();
+        cloudAPI.getService().setServiceState(ServiceState.RUNNING_UNDEFINED);
+        cloudAPI.getService().setOnlineCount(ProxyServer.getInstance().getOnlineCount());
+        cloudAPI.getService().update();
+    }
+
+    @Override
+    public void onDisable() {
+        cloudAPI.shutdown(false);
+    }
+}
