+++ plugins/plugin-bungeecord/src/main/java/net/suqatri/redicloud/plugin/bungeecord/listener/PostLoginListener.java
@@ -1,0 +1,18 @@
+package net.suqatri.redicloud.plugin.bungeecord.listener;
+
+import net.md_5.bungee.api.event.PostLoginEvent;
+import net.md_5.bungee.api.plugin.Listener;
+import net.md_5.bungee.event.EventHandler;
+import net.suqatri.redicloud.plugin.bungeecord.BungeeCordCloudAPI;
+
+public class PostLoginListener implements Listener {
+
+    @EventHandler
+    public void onPostLogin(PostLoginEvent event){
+        if(BungeeCordCloudAPI.getInstance().getService().isMaintenance() && !event.getPlayer().hasPermission("redicloud.maintenance.bypass")){
+            event.getPlayer().disconnect("§cThis proxy is currently under maintenance. Please try again later.");
+            return;
+        }
+    }
+
+}
