+++ plugins/plugin-bungeecord/src/main/resources/bungee.yml
@@ -1,0 +1,4 @@
+name: RediCloud-Plugin
+author: Suqatri
+version: 1.0-SNAPSHOT
+main: net.suqatri.redicloud.plugin.bungeecord.BungeeCordCloudPlugin
