+++ plugins/plugin-minecraft/build.gradle
@@ -2,7 +2,7 @@ plugins {
     id 'java'
 }
 
-group 'net.suqatri.redicloud'
+group 'net.suqatri.redicloud.plugin.minecraft'
 
 dependencies {
     implementation project(':commons')
